+++ src/components/Modals/YouTubeIFrame/index.tsx
@@ -1,18 +1,16 @@
-import {
-  Feather,
-  MaterialCommunityIcons,
-  MaterialIcons,
-} from "@expo/vector-icons";
+import { Feather, MaterialCommunityIcons } from "@expo/vector-icons";
+import { secondsToMilliseconds } from "date-fns";
+import { AnimatePresence, MotiView } from "moti";
 import React, {
   forwardRef,
-  memo,
   useCallback,
   useEffect,
   useImperativeHandle,
-  useMemo,
   useRef,
   useState,
 } from "react";
+import { StatusBar } from "react-native";
+import SystemNavigationBar from "react-native-system-navigation-bar";
 import { useTheme } from "styled-components";
 import { millisToTime } from "../../../utils/format";
 import YouTubeVideoPlayer, {
@@ -43,16 +41,21 @@ export interface IYouTubeIFrameRef {
   openYouTubeIFrameModal: () => void;
 }
 
-const YouTubeIFrame: React.ForwardRefRenderFunction<IYouTubeIFrameRef, IYouTubeIFrame> = (
-  { videoId },
-  ref
-) => {
+const TIME_FOR_HIDE_CONTROLS = secondsToMilliseconds(6)
+
+const YouTubeIFrame: React.ForwardRefRenderFunction<
+  IYouTubeIFrameRef,
+  IYouTubeIFrame
+> = ({ videoId }, ref) => {
   const [videoStatus, setVideoStatus] = useState<"PLAYING" | "PAUSED">(
     "PLAYING"
   );
   const [videoDuration, setVideoDuration] = useState(0);
   const [modalVisible, setModalVisible] = useState(false);
   const [currentTime, setCurrentTime] = useState(0);
+  const [hiddenControls, setHiddenControls] = useState(true);
+  const [hiddenControlsTimeout, setHiddenControlsTimeout] =
+    useState<NodeJS.Timeout>();
   const ytPlayerRef = useRef<IYouTubeControllers>(null);
 
   const { colors } = useTheme();
@@ -67,9 +70,12 @@ const YouTubeIFrame: React.ForwardRefRenderFunction<IYouTubeIFrameRef, IYouTubeI
     setVideoStatus((old) => (old === "PLAYING" ? "PAUSED" : "PLAYING"));
   }, [ytPlayerRef, videoStatus]);
 
-  const seekTo = useCallback((time: number) => {
-    ytPlayerRef.current.seekTo(time);
-  }, []);
+  const seekTo = useCallback(
+    (time: number) => {
+      ytPlayerRef.current.seekTo(time);
+    },
+    [ytPlayerRef]
+  );
 
   const onUpdateTime = useCallback((time: number) => {
     setCurrentTime(time);
@@ -78,6 +84,26 @@ const YouTubeIFrame: React.ForwardRefRenderFunction<IYouTubeIFrameRef, IYouTubeI
   const openYouTubeIFrameModal = useCallback(() => {
     setModalVisible(true);
   }, [modalVisible]);
+
+  const handleCloseModal = useCallback(async () => {
+    await SystemNavigationBar.navigationShow();
+    setModalVisible(false);
+  }, [modalVisible]);
+
+  const hideAndShowControls = useCallback(async () => {
+    if (!hiddenControls) {
+      setHiddenControls(true);
+      return;
+    }
+
+    setHiddenControls(false);
+
+    const timeout = setTimeout(async () => {
+      setHiddenControls(true);
+    }, TIME_FOR_HIDE_CONTROLS);
+    clearTimeout(hiddenControlsTimeout);
+    setHiddenControlsTimeout(timeout);
+  }, [hiddenControls, hiddenControlsTimeout]);
 
   useEffect(() => {
     if (ytPlayerRef.current) {
@@ -92,23 +118,22 @@ const YouTubeIFrame: React.ForwardRefRenderFunction<IYouTubeIFrameRef, IYouTubeI
   return (
     <Container
       visible={modalVisible}
-      animationType="slide"
+      onRequestClose={handleCloseModal}
+      supportedOrientations={["landscape", "portrait"]}
       statusBarTranslucent
-      onDismiss={() => setModalVisible(false)}
     >
-      <YouTubeModalHeader>
-        <YouTubeModalHeaderButton>
-          <Feather
-            onPress={() => setModalVisible(false)}
-            name="x"
-            size={25}
-            color="#fff"
-          />
-        </YouTubeModalHeaderButton>
-        <YouTubeModalHeaderButton>
-          <MaterialCommunityIcons name="fullscreen" size={28} color="#fff" />
-        </YouTubeModalHeaderButton>
-      </YouTubeModalHeader>
+      {!hiddenControls && (
+        <YouTubeModalHeader
+          from={{ opacity: 0, translateY: -50 }}
+          animate={{ opacity: 1, translateY: 0 }}
+          exit={{ opacity: 0 }}
+          transition={{ type: "timing", duration: 300 }}
+        >
+          <YouTubeModalHeaderButton onPress={handleCloseModal}>
+            <Feather name="x" size={25} color="#fff" />
+          </YouTubeModalHeaderButton>
+        </YouTubeModalHeader>
+      )}
       <YouTubeModal>
         <YouTubeVideoPlayer
           ref={ytPlayerRef}
@@ -116,46 +141,64 @@ const YouTubeIFrame: React.ForwardRefRenderFunction<IYouTubeIFrameRef, IYouTubeI
           onUpdateTime={onUpdateTime}
           autoplay
         />
-        <YouTubePlayerControlsContainer>
-          <YouTubePlayerControls>
-            <YouTubePlayerPlayAndPauseContainer>
-              <YouTubePlayerPlayAndPauseButton
-                activeOpacity={0.7}
-                onPress={playPauseVideo}
+        <YouTubePlayerControlsContainer onPress={hideAndShowControls}>
+          <AnimatePresence>
+            {!hiddenControls && (
+              <MotiView
+                from={{ opacity: 0 }}
+                animate={{ opacity: 1, scale: 1 }}
+                exit={{
+                  opacity: 0,
+                  scale: 1.1,
+                }}
+                transition={{
+                  type: "timing",
+                  duration: 200,
+                }}
+                style={{ flex: 1 }}
               >
-                <MaterialCommunityIcons
-                  name={videoStatus === "PLAYING" ? "pause" : "play"}
-                  size={33}
-                  color="#fff"
-                />
-              </YouTubePlayerPlayAndPauseButton>
-            </YouTubePlayerPlayAndPauseContainer>
-            <YouTubePlayerInfosContainer>
-              <YouTubePlayerInfos>
-                <YouTubePlayerInfoContainer>
-                  <YouTubePlayerInfoText>
-                    {millisToTime(currentTime * 1000)}
-                  </YouTubePlayerInfoText>
-                </YouTubePlayerInfoContainer>
-                <YouTubePlayerInfoContainer>
-                  <YouTubePlayerInfoText>
-                    {millisToTime(videoDuration * 1000)}
-                  </YouTubePlayerInfoText>
-                </YouTubePlayerInfoContainer>
-              </YouTubePlayerInfos>
-            </YouTubePlayerInfosContainer>
-            <YouTubePlayerSeekBarContainer>
-              <YouTubePlayerSeekBar
-                value={currentTime}
-                minimumValue={0}
-                maximumValue={videoDuration}
-                thumbTintColor={colors.secondary}
-                minimumTrackTintColor={colors.secondary}
-                maximumTrackTintColor={colors.dark_gray}
-                onSlidingComplete={seekTo}
-              />
-            </YouTubePlayerSeekBarContainer>
-          </YouTubePlayerControls>
+                <YouTubePlayerControls>
+                  <YouTubePlayerPlayAndPauseContainer>
+                    <YouTubePlayerPlayAndPauseButton
+                      activeOpacity={0.7}
+                      onPress={playPauseVideo}
+                    >
+                      <MaterialCommunityIcons
+                        name={videoStatus === "PLAYING" ? "pause" : "play"}
+                        size={33}
+                        color="#fff"
+                      />
+                    </YouTubePlayerPlayAndPauseButton>
+                  </YouTubePlayerPlayAndPauseContainer>
+                  <YouTubePlayerInfosContainer>
+                    <YouTubePlayerInfos>
+                      <YouTubePlayerInfoContainer>
+                        <YouTubePlayerInfoText>
+                          {millisToTime(currentTime * 1000)}
+                        </YouTubePlayerInfoText>
+                      </YouTubePlayerInfoContainer>
+                      <YouTubePlayerInfoContainer>
+                        <YouTubePlayerInfoText>
+                          {millisToTime(videoDuration * 1000)}
+                        </YouTubePlayerInfoText>
+                      </YouTubePlayerInfoContainer>
+                    </YouTubePlayerInfos>
+                  </YouTubePlayerInfosContainer>
+                  <YouTubePlayerSeekBarContainer>
+                    <YouTubePlayerSeekBar
+                      value={currentTime}
+                      minimumValue={0}
+                      maximumValue={videoDuration}
+                      thumbTintColor={colors.secondary}
+                      minimumTrackTintColor={colors.secondary}
+                      maximumTrackTintColor={colors.dark_gray}
+                      onSlidingComplete={seekTo}
+                    />
+                  </YouTubePlayerSeekBarContainer>
+                </YouTubePlayerControls>
+              </MotiView>
+            )}
+          </AnimatePresence>
         </YouTubePlayerControlsContainer>
       </YouTubeModal>
     </Container>