+++ src/utils/index.ts
@@ -8,3 +8,7 @@ export function verifyBetweenValues(
 
   return value > min && value < max
 }
+
+export function convertSecondsToMs(seconds: number) {
+  return seconds * 1000;
+}